LPS-146894 Move code to new module
