+++ modules/apps/words/words-impl/src/main/resources/com/liferay/words/internal/dependencies/centre.dic
