+++ modules/apps/words/words-impl/src/main/resources/com/liferay/words/internal/dependencies/ise.dic
